refactor: Tax 관련 타입을 types/tax.ts로 분리
