+++ src/lib/calculateTax.ts
@@ -1,14 +1,4 @@
-export interface TaxInput {
-  income: number; // 연소득
-  expense: number; // 필요경비
-}
-
-export interface TaxResult {
-  taxableIncome: number; // 과세표준
-  taxAmount: number; // 산출세액
-  appliedRate: number; // 적용 세율(%)
-  deduction: number; // 누진공제
-}
+import { TaxInput, TaxResult } from "@/types/tax";
 
 export default function calculateTax({ income, expense }: TaxInput): TaxResult {
   const BASIC_DEDUCTION = 1_500_000; // 기본 공제액(150만원)
